tag 1.2.4; fix software rendering for crop offsets
Affects Theora frames at least, eg at 854x480

The dx and dy parameters of putImageData don't do
what you't expect in combination with the dirty
rect params. Well it makes sense when you think
of it as a dirty rect, but not when you think of
it as a source rect. :)

Must offset the offsets to place the image in the
correct place.