1.2.10 - re-disable stripe with fixes
Part of the reason the stripe mode was fast was that uploads
were not using gl.texSubImage2d on the non-stripe code path
when updating an existing texture.

This is still slower on macOS 12.3 than with the stripe mode
off, but is probably fast enough.

Note that Firefox has some troubles with the stripe mode on
macOS at present -- I'm not sure why, but it seems to be
something with the framebuffer used to unpack textures.

Anyway, leaving it off by default resolves functionality
at a modest cost of performance on current macOS/iOS and
I may remove the stripe code later, but am for now leaving
it in so it can be re-enabled manually if needed.

Resolves #37